Moved the core o the action system out of the game package since it is used equally in the client. The ActionUnion is moved into the core module
