+++ console-client/src/main/java/machinestrike/client/console/ConsoleClient.java
@@ -12,8 +12,8 @@ import machinestrike.debug.Assert;
 import machinestrike.game.Game;
 import machinestrike.game.Orientation;
 import machinestrike.game.Player;
-import machinestrike.game.action.Action;
-import machinestrike.game.action.GameActionHandler;
+import machinestrike.action.Action;
+import machinestrike.action.GameActionHandler;
 import machinestrike.game.level.Board;
 import machinestrike.game.level.factory.BoardFactory;
 import machinestrike.game.level.factory.DefaultBoardFactory;
@@ -25,7 +25,7 @@ import machinestrike.game.rule.RuleBook;
 import machinestrike.game.rule.RuleViolation;
 import machinestrike.game.rule.factory.DefaultRuleBookFactory;
 import machinestrike.game.rule.factory.RuleBookFactory;
-import machinestrike.util.ActionUnion;
+import machinestrike.action.ActionUnion;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
