+++ console-client/src/main/java/machinestrike/client/console/action/ConsoleAction.java
@@ -1,6 +1,6 @@
 package machinestrike.client.console.action;
 
-import machinestrike.game.action.Action;
+import machinestrike.action.Action;
 
 public interface ConsoleAction extends Action<ConsoleActionHandler> {
 }
