+++ console-client/src/main/java/machinestrike/client/console/input/InputHandler.java
@@ -1,8 +1,8 @@
 package machinestrike.client.console.input;
 
-import machinestrike.util.ActionUnion;
+import machinestrike.action.ActionUnion;
 import machinestrike.client.console.action.ConsoleActionHandler;
-import machinestrike.game.action.GameActionHandler;
+import machinestrike.action.GameActionHandler;
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
