+++ console-client/src/main/java/machinestrike/client/console/input/parser/SprintActionParser.java
@@ -1,8 +1,8 @@
 package machinestrike.client.console.input.parser;
 
-import machinestrike.util.ActionUnion;
+import machinestrike.action.ActionUnion;
 import machinestrike.client.console.action.ConsoleActionHandler;
-import machinestrike.game.action.GameActionHandler;
+import machinestrike.action.GameActionHandler;
 import machinestrike.game.action.MoveAction;
 import org.jetbrains.annotations.NotNull;
