+++ core/src/main/java/machinestrike/action/GameActionHandler.java
@@ -1,5 +1,7 @@
-package machinestrike.game.action;
+package machinestrike.action;
 
+import machinestrike.game.action.AttackAction;
+import machinestrike.game.action.MoveAction;
 import machinestrike.game.rule.RuleViolation;
 import org.jetbrains.annotations.NotNull;
