+++ core/src/main/java/machinestrike/game/Game.java
@@ -1,9 +1,9 @@
 package machinestrike.game;
 
 import machinestrike.debug.Assert;
-import machinestrike.game.action.Action;
+import machinestrike.action.Action;
 import machinestrike.game.action.AttackAction;
-import machinestrike.game.action.GameActionHandler;
+import machinestrike.action.GameActionHandler;
 import machinestrike.game.action.MoveAction;
 import machinestrike.game.level.Board;
 import machinestrike.game.machine.Machine;
