+++ core/src/main/java/machinestrike/game/action/GameAction.java
@@ -1,4 +1,7 @@
 package machinestrike.game.action;
 
+import machinestrike.action.Action;
+import machinestrike.action.GameActionHandler;
+
 public interface GameAction extends Action<GameActionHandler> {
 }
