+++ core/src/main/java/machinestrike/game/action/MoveAction.java
@@ -1,5 +1,6 @@
 package machinestrike.game.action;
 
+import machinestrike.action.GameActionHandler;
 import machinestrike.game.Orientation;
 import machinestrike.game.Point;
 import machinestrike.game.rule.RuleViolation;
